Add check for edge cases and add prisma disconnections

+++ pages/api/user/[id].ts
@@ -1,5 +1,6 @@
 import { PrismaClient, User } from '@prisma/client'
 import type { NextApiRequest, NextApiResponse } from 'next'
+import { isEmailUsed } from '../../../helpers/api/validate'
 
 
 export default async function handler(
@@ -7,17 +8,19 @@ export default async function handler(
 	res: NextApiResponse<User>
 ) {
 	const id = Number(req.query.id)
-	const body = JSON.parse(req.body)
-
 	const prisma = new PrismaClient()
 	const user = await prisma.user.findUnique({
 		where: {
 			id: id
+		},
+		include: {
+			tank: true
 		}
 	})
 
 	if (!user) {
 		res.status(404).end(`User With ID ${id} Not Found`)
+		await prisma.$disconnect()
 		return
 	}
 
@@ -25,8 +28,13 @@ export default async function handler(
 		case 'GET':
 			res.status(200).json(user)
 			break
+
 		case 'PUT':
-			// TODO: fail if email already being used
+			const body = JSON.parse(req.body)
+			if (await isEmailUsed(body.email)) {
+				res.status(409).end(`The Email ${body.email} Is Already Used`)
+			}
+
 			const updatedUser = await prisma.user.update({
 				where: {
 					id: id
@@ -38,17 +46,24 @@ export default async function handler(
 			})
 			res.status(200).json(updatedUser)
 			break
+
 		case 'DELETE':
-			// TODO: fail if foreign key
+			if (user.tank) {
+				res.status(409).end("The User Is Linked To A Tank")
+				await prisma.$disconnect()
+				return
+			}
 			const deletedUser = await prisma.user.delete({
-			where: {
-				id: id
+				where: {
+					id: id
 				}
 			})
 			res.status(200).json(deletedUser)
 			break
 		default:
 			res.status(405).end(`Method ${req.method} Not Allowed`)
-			break
+			await prisma.$disconnect()
+			return
 	}
+	await prisma.$disconnect()
 }
